refactoring shraga client
