+++ src/main/java/com/cx/restclient/dto/IResults.java
@@ -1,0 +1,4 @@
+package com.cx.restclient.dto;
+
+public interface IResults {
+}
